+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### Bug Fixes:
 
+- fix(auth): `fastly profile`, `fastly sso` and `fastly auth-token` commands now correctly respect the `--quiet` flag [#1706](https://github.com/fastly/cli/pull/1706)
+
+
 ### Enhancements:
 
 - feat(vcl/snippet): add support for the '--content' flag, allowing for the raw output of VCL.  [#1706](https://github.com/fastly/cli/pull/1706)